+++ aiohttp_sqlalchemy/middlewares.py
@@ -12,12 +12,23 @@ def sa_middleware(key: str = SA_DEFAULT_KEY) -> THandler:
     """
 
     @middleware
-    async def sa_middleware_(request: Request, handler: THandler) -> StreamResponse:
+    async def sa_middleware_(
+        request: Request,
+        handler: THandler,
+    ) -> StreamResponse:
         if key in request:
             raise DuplicateRequestKeyError(key)
 
+        # TODO: after dropped Python 3.7
+        # if session_factory := request.config_dict.get(key):
         session_factory = request.config_dict.get(key)
-        async with session_factory() as request[key]:
-            return await handler(request)
+        if session_factory:
+            async with session_factory() as request[key]:
+                return await handler(request)
+        else:
+            raise KeyError(
+                f'Session factory not found by {key}.'
+                'Check `aiohttp_sqlalchemy.setup()`.'
+            )
 
     return sa_middleware_