+++ tests/test_setup.py
@@ -7,31 +7,22 @@ from aiohttp_sqlalchemy.typedefs import TSessionFactory
 
 
 async def test_duplicate_app_key_error(
-    orm_session_factory: TSessionFactory,
+    session_factory: TSessionFactory,
 ) -> None:
 
     with pytest.raises(DuplicateAppKeyError):
-        aiohttp_sqlalchemy.setup(
-            web.Application(),
-            [
-                aiohttp_sqlalchemy.bind(orm_session_factory),
-                aiohttp_sqlalchemy.bind(orm_session_factory),
-            ],
-        )
+        aiohttp_sqlalchemy.setup(web.Application(), [
+            aiohttp_sqlalchemy.bind(session_factory),
+            aiohttp_sqlalchemy.bind(session_factory),
+        ])
 
     with pytest.raises(DuplicateAppKeyError):
-        aiohttp_sqlalchemy.setup(
-            web.Application(),
-            [
-                aiohttp_sqlalchemy.bind(orm_session_factory, "sa_secondary"),
-                aiohttp_sqlalchemy.bind(orm_session_factory, "sa_secondary"),
-            ],
-        )
+        aiohttp_sqlalchemy.setup(web.Application(), [
+            aiohttp_sqlalchemy.bind(session_factory, 'sa_secondary'),
+            aiohttp_sqlalchemy.bind(session_factory, 'sa_secondary'),
+        ])
 
-    aiohttp_sqlalchemy.setup(
-        web.Application(),
-        [
-            aiohttp_sqlalchemy.bind(orm_session_factory),
-            aiohttp_sqlalchemy.bind(orm_session_factory, "sa_secondary"),
-        ],
-    )
+    aiohttp_sqlalchemy.setup(web.Application(), [
+        aiohttp_sqlalchemy.bind(session_factory),
+        aiohttp_sqlalchemy.bind(session_factory, 'sa_secondary'),
+    ])